MakerArm: Add SCARA debug flags

+++ Marlin/Marlin.h
@@ -189,7 +189,7 @@ void kill(const char*);
 
 void quickstop_stepper();
 
-extern uint8_t marlin_debug_flags;
+extern uint16_t marlin_debug_flags;
 #define DEBUGGING(F) (marlin_debug_flags & (DEBUG_## F))
 
 extern bool Running;
